requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pymupdf
openpyxl
xlrd
celery[redis]
flower
redis
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-magic
httpx
requests
slowapi
reportlab
pydantic-settings
deep-translator
python-docx
pytesseract
